LWIS: vmdir provider bindprotocol: POST REST Token authorization
Bug Number: 2030254

Changing the default bind protocol used by lsass vmdir provider from
Kerberos to SRP. We faced issues when the protocol was kerberos because
the configuration between POST and lightwave was not accurate for
Kerberos.

Tests:

Ran the two commands manually to ensure that the protocol is set
properly.

Change-Id: I53b9b09580ea4c56f01032af1f2a0c165c702140