POST returning weird error while creating Cluster
Bug Number: 2029510

This change was made to fix the error translation in POST. For a case
where a node switches role from being a leader->follower after a rest
transaction is performed but before it is committed, mdb layer returns a
VMDIR_ERROR 9114 (UNWILLING_TO_PERFORM) which should have been ideally
mapped to 503 (Service Unavailable) which is retriable instead of 500
(Internal Server Error)

The reason for the incorrect mapping was while translating an error from
MDB to VMDIR error base, we always return BACKEND_ERROR. The fix checks
if the error exists in the VMDIR error base. If it does it is not
chnaged to BACKEND_ERROR. As a part of the change we also mapped
LDAP_UNWILLING_TO_PERFORM to 503.

Change-Id: I5211e0a7f5cc981b06c3ac54b8d981f1ed54ea3b