Use only PRIV_USER (root) ssh creds during promotion
Got rid of UNPRIV_USER, as the principal of "least privilege" only
complicates AD promotion in this case, as root must be used to install
RPMS, and perform other operations.

Modified path for SASL to use designated LIGHTWAE_BASE/vmdir/thirdparty, as
that is where the modified CyrusSASL is currently checked in to git.

Change-Id: I3a119f54aa6534bfa82e3fc721669d7bbfdbffd1