core management in AWS: upload core file and build info to S3 bucket for LW services (Part 1)
Bug Number: 2033735

This change was made to add a cronjob which runs every hour and collects
coredump files if any and copies them over to a S3 bucket in AWS. The
coredump is saved under a folder name corresponding to the build number.

Test Done:

In unstable post manually ran the cronjob script. Was able to upload the
test files successfully onto S3. The job is scheduled to run every hour 
I made sure that each ASG has the required tag values and S3 buckets.

Change-Id: Iab91602f99585b50920f50d4efa42a9f5f02573d