CI/CD Stable/BVT Blocker: BVT deployments are failing due error "Cannot communicate with object store"
Bug Number: 2036042

This change was made to add logging when a token gets presented to post.
We log if the upn belongs to lightwave admin group and hence is given
admin access rights on post. This change will help us track down the
intermittent issue in the bvt where an add request fails with invalid
access rights error.

Change-Id: Ia18d1272a97c481b720b03bd55a51adae008d2f9