Fix CLDAP response username
Fixed CLAP response username to be the NetBIOS name of the joining client, when
provided in the search filter.
Query and return the actual domain GUID value, vs an incorrect hard-coded value.
Changed result to return an empty dn: value for only DSE Root queries.
Set file permissions on replacement SASL libraries to be world-readale.

Change-Id: I97b5dc31d61a7cf55dd33e620fba07302aae87ba