[25x20] Promote POST error=1 (rpc_s_connect_timed_out) Connection request timed out 382312513
Bug Number: 2053218

This change was made to address a bug in the iptables configuration in
POST. Currently after iptables refresh RPC endpoint was not opened as a
result of which the cluster got in unrecoverable state. This was because
the RPC endpoint was missing in the firewall configuration for POST.

I have added the endpoint for POST and VMDIR. All other services already have
their respective RPC ports listed in the firewall config.

Test:

Installed on dev box and made sure the iptables list the right ports.

Change-Id: I7dfa2f397484cd650d4e7370e7dcc37c607474c8