Changing the refresh network services cron job to run when aws fails
This change includes the fix for the issue when POST has an invalid DNS
server in the /etc/resolv.conf but cannot recover because the refresh
job talk to AWS and keeps failing.
In this change I added code to check if any of the AWS calls fail. If
they do, we force refresh.

Test:
Ran locally where all aws calls fail, the network services are
restarted.

Change-Id: Ic45b758fd157a1db54012ade912ad6c4f20d692c