Add site parameter to credentials validation API
This change adds an optional site parameter to
VmAfdJoinValidateDomainCredentialsA().  This fixes an
issue where long delays can occur when trying to
connect to DCs in inaccessible sites.

Testing: promoted two DCs for a domain in separate sites.
Joined a client to each site, the correct DC was selected
for validating creds.
PR#2065267

Change-Id: I71469f5541f2fd2f2b6096f107761ef4291846fb