lock around ldap calls in RootFetchTask
There appears to be a reentrancy issue in the SASL
library when ldap client connections are made from multiple
threads.  This change expands the section of code in
RootFetchTask that is protected by the lock to include
the ldap calls.

Testing: Ran ic-join with --atomic flag while vmafdd ran under
valgrind. Before the change valgrind reported invalid reads
and writes from a block of memory that had been freed by
the SASL library.  After this change valgrind did not report
and invalid reads or writes.

PR#2072309

Change-Id: I32ef064737da25d48a8f4f241640540c0a769d1b