Added utility to verify if two DBs contain the same data.
mdb_verify_checksum takes one or two DB paths as input. If two DBs
are given it will calculate the running SHA sum of each key/value
in each DB and compare the two sums in the end.
If only one DB is given as input it will print the checksum of that DB.
For further debugging it can also calculate/compare the checkum of
a particular subDB using the -s option.
Added compilation of mdb_stat. This will be needed later during compaction.
Modified the rpm spec to copy the two binaries during rpm install.

Change-Id: Id6c6734fa440ae6f468f7ee4711b90ddf2a2c658