VMDIR support for domain trusts
This change adds new APIs to VMDIR for supporting
hierarchical domain trusts:

  VmDirSetupHostInstanceTrust: This sets up a new
     host instance as a child of an existing
     domain.  Shared keys are created for Kerberos
     cross-realm authentication in a trustedDomain
     object within the System container.

  VmDirCreateDomainTrust: This creates a domain
     trust with shared keys.

The KDC now handles requests for Kerberos cross-realm
keys by looking up the appropriate trustedDomain objects.
If a direct trust is not available, it may return a
ticket for a server closer in the hierarchical path to
allow walking the realm tree.

Change-Id: If17c5ca0a9fef25fc479c7870216186bc9ec3e69