Preserve UPN casing on user JIT creation
Before, during the CSP tenant and user JIT creation, UPN casing was not
preserved, as the domain portion was converted to upper case. This
change preserves the casing when the user is added to directory.

Changes:
- Use TreeSet with Case insensitive comparator for storing upn suffixes
so contains method is case insensitive

Testing:
- Ran OIDC integration test and rest/idm/client integration test
- Verified manually: Registered upn suffixes with alternating case,
created users with alternating case, verified using /finder/principals
api that upn casing was preserved from argument to create user

Change-Id: I1d46569117f108f9aecd3dddbca6f3ce63dfadef