Delete DNS records on domain cleanup
This change removes PTR records when uninitializing
a domain.  When A or AAAA records are removed, any PTR
records that correspond with their IP addresses are also
removed.

Also, when a DNS record is deleted, and the LDAP object
has no more dnsRecord attributes, remove the LDAP object.

Testing:
  1) promote standalone dc
  2) promote partner dc, this creates A and PTR records
  3) demote partner dc
The A and PTR records for the partner dc were removed and
the LDAP objects were deleted.

Change-Id: I67c260614551e465a4dd17cb1a3a1cda5c65ec22