Override OIDC Endpoints With Localhost If Local
As per OIDC standards, our OIDC client libraries that the various REST
heads for our services use query the OIDC configuration and then use the
URL's returned to communicate with IDM.

As an optimization, now instead of going to the URL specified in the
OIDC configuration every time, we go to the local IDM instance if the
source address is localhost.

Changes:

    * vmidentity/ssoclients/common/include/common.h - define new string
utility (sprintf) and function to determine if localhost
    * vmidentity/ssoclients/common/src/Makefile.am - build util.c which
has functions to determine if localhost
    * vmidentity/ssoclients/common/src/includes.h - include libraries
necessary for util.h's localhost check function
    * vmidentity/ssoclients/common/src/strings.c - string function for
sprintf
    * vmidentity/ssoclients/common/util.c - public API to determine if
an address is localhost and helper functions for this
    * vmidentity/ssoclients/oidc/src/oidc_server_metadata.c - added
logic to OidcServerMetadataAcquire to check if requesting server is
localhost and then override the OIDC configuration values for token and
JWKS URIs.

Testing:

    * Submitted REST request to VMDIR and tracked through GDB which URL
is used to talk to OIDC (localhost vs config's value) and it was
localhost
    * Grabbed PCAP on host and ensured that communication from VMDIR to
OIDC was through localhost and not the OIDC config specified value

Change-Id: Ic5f3aef19ca90ccd52b9aaa7b26b3af5de529a14