Encode tenant before returning SAML response
chooseidp.jsp displays the tenant from what is returned from the SAML
request. If the SAML request does not properly validate tenant, there is
a chance for XSS vulnerability. This change ensures the tenant paramenter
derived from path is encoded before returning. SAML requests involving
tenant as a parameter will also validate the tenant in IDM before adding
it as an attribute to the spring model.

Bug Number: 2080715

Change-Id: I08de31ab63a775cc2f25110c7841f1344fa7c694