Lightwave AWS schema patch script
We will run patch/patch-schema.sh to perform schema patch in a separate
action before the in-place upgrade action, so that we won't initiate the
in-place upgrade if schema patch can't pass.

The schema patch tool is going to do a dryrun first to validate the
patch complies the schema and is backward compatible.

Jira ID: CNA-776

Change-Id: Ib17e046a1e66aa69e788203e4a755b4bc83a9e4a