Enable CDC thread on DCs and add native interface
As part of JIT resilience work, IDM will create tenant on an instance
that is determined by CDC to be healthy. This change enables DCs to ping
partners for heartbeat, and adds functionality for IDM to get the status
of all partner DCs. With the status info for all DCs, IDM can
deterministically use one for remote tenant creations.

Changes:
- modify cdc to allow domain controllers to run caching background
thread but not state machine
- Add java interop to call GetDCStatusInfoW
- Add functionality to enumerate all partner DCs and get status

Testing:
- added tests in interop and verified it passed
- ran integration tests
- manual test: on 2 node instance, enumerated partner DCs and printed
all fields, verified individual service health was printed along with DC
info

Change-Id: Ia7152704301229900d21f3dcc34fb30770c14077