Fix mode argument to mdb_env_open()
VMDIR and POST were calling mdb_env_open() with
the mode argument set to O_RDWR.  This is incorrect,
passing an open flags value as a mode parameter.

This change renames the argument from "oflags" to "mode"
to better reflect the meaning of the argument, and sets
the value to "S_IRUSR | S_IWUSR" which evaluates to mode
0600 (user read and write).

Prior to this change the permissions on the MDB files
were set to mode 0000 (no permission bits set) because
the incorrect mode O_RDWR (2 - other write) was masked
by the umask 022.

Vmdir MDB file permissions before this change:

total 8612
---------- 1 root root 8814592 May  8 22:35 data.mdb
---------- 1 root root   64128 May  8 22:35 lock.mdb

Since VMDIR and POST run as root they could still open
these files.  In the future when the processes are
run as a non-root user, this fix will be necessary.

Vmdir MDB file permissions after this change:

total 8612
-rw------- 1 root root 8814592 May  8 23:01 data.mdb
-rw------- 1 root root   64128 May  8 23:01 lock.mdb

Change-Id: I64d5ab359ca1b5945b9443edfa567ac2ea781fb2