Go OIDC Lib: Add APIs to Get HOTK Claim and PubKey
This change adds APIs to get the public key from the HOTK claim if the
token is of type hotk-pk.

It adds to the token validation function to ensure that if the token
type is hotk-pk, it will verify that there is a HOTK claim present.

It adds the new API, Hotk() which will retreive the HOTK public key i
form of *rsa.PublicKey.

This change also adds unit tests to verify that the API works as
intended.

Testing:

    * Unit tests to verify library and new API functionality

Change-Id: I028fb991af2e9fe33da5ae3951c41876ae87ed43