Ignore error on Post vmafd heartbeat
When a client calls StartHeartbeat() and vmafd is not running, the start
heartbeat call will fail. This change ignores errors on the first
PostHeartbeat() to vmafd to allow the heartbeat worker thread to run.
This way, if vmafd runs after StartHeartbeat call, the heartbeats from
the services will be registered.

Changes:
- Ignore errors on first PostHeartbeat when starting heartbeat

Testing:
- started vmdir before vmafd, waited, and started vmafd. Verified
"Was not able to start vmafd heartbeat" not in logs and heartbeat status
was still reporting 1 after 2 minutes.

Change-Id: I41c894f271b5efcd9015d8d98993c238f9e0bee4