Temporary change in DR script to handle LW node upgrade
Scale up to two nodes before upgrading to latest RPMs instead of deleting the deployment group.
This increases DR time by about 15 mins for now. Once the new LW pipeline is in place,
this can be removed and single node upgrade should work.
Also added check to delete codedeploy hook only if not null. So there is no need to reinint the hook
if the script fails and needs to be rerun.
Updated usage.

Change-Id: If325e6da182237ca2e2a5bfe90d13781cb79b60c