Enforce hotk POP Verification In VMDIR/POST/VMCA
Before, requests to the VMDIR/POST/VMCA REST head that presented an
hotk-pk token and proof of posession (POP) signature would be
authenticated, even if the POP could not be verified.  This was done for
debugging reasons to ensure that various clients could generate
verifiable POPs.

We are at the point where we can disable this debugging pass-through.
Now, POP verification is enforced and if a POP cannot be verified, then
the request will be failed and the appropriate HTTP status code will be
returned.

Testing:

    * Sent REST commands to each VMDIR/POST/VMCA with Bearer token --
    authenticated
    * Sent REST commands to each VMDIR/POST/VMCA with hotk-pk token
    without POP -- failed to authenticate
    * Sent REST commands to each VMDIR/POST/VMCA with hotk-pk and
    invalid POP -- failed to authenticate
    * Sent REST commands to each VMDIR/POST/VMCA with correct POP --
    authenticated
    * Sent REST commands to each VMDIR/POST/VMCA with tampered request
    body (request body must be same input used to create POP) -- failed
    to authenticate

Change-Id: I61cd2c0a141c5067c654d4238639d0b5ca730fa6