Fix introduced bug where objectGUID are always returned
The implementation of binary SID / GUID did not check the
search request attibute list for "objectSID" and "objectGUID".
These values were included in the LDAP response should these
attributes were present in the actual object searched in the directory.
The result was when "objectSID" or "objectGUID" was in the searched
directory objet, these values were always returned in the search result.

This was causing erronious repeated LDAP search requests from
Windows during Domain Join to Lightwave.

Change-Id: Ica6f40c623aba2b31a44fb37797eee268e57a09e