New: POST: wavefront alert: REST API: should ignore 9300 benign error code to prevent false alert
Bug Number: 2133436

Vmdir error 9300 is a custom error code and hence is not mapped to any
of the known LDAP error codes. Instead gets mapped to LDAP_OTHER. As a
part of this change the error code will be handled seperately while
sending metrics. We will change the alerts on wavefront to ignore this
error code while triggering an alert.

Change-Id: I1f430d088bfff2da482ae519f910e50064710bc1