Relax VECS ForceRefresh Permissoins and Add Go API
Before, only the root user could call vecs-cli force-refresh command.
Since we have deprivleged Lightwave, now, we need members of the group,
lightwave, to be able to run this vecs-cli command.

This change addresses that by changing the privilege check in the
`VmAfdIpcTriggerRootCertsRefresh` function to lookup the group
membership set in the registry and then check privileges of the caller
against that.

It also adds two client API to the VECS Golang interop:

    * `VecsForceRefreshCerts` trigger a force refresh of certificates.
    * `VecsGetEntryCount` which returns the number of entries in a VECS
    store

Bug Number: 2133525

Testing:

    * Ran vecs-cli force-refresh command as both root and non-root user
    * Wrote a test program in Golang which calls VecsForceRefreshCerts
    API
    * Manually deleted all Lightwave Trusted Root certs from the
    `/etc/ssl/certs` directory
    * Ran Go test program as root and then searched the `/etc/ssl/certs`
    directory and found the Lightwave Trusted Root certs again
    * Created a non-root user and added it to the lightwave group.
    Deleted the Lightwave Trusted Root certs from `/etc/ssl/certs` and
    ran Go test program.  Querying `/etc/ssl/certs` again resulted in
    the certs being present
    * Unit tests for VecsGetEntryCount API

Change-Id: I011998c953ed6cbd5e69979b4defd9ceca217d3e