Added LDAP file copy functionality
Added a new LDAP control which accepts the file name/fd and blocksize
and returns data read from the file.  Implemented logic to use this
new functionality in the new node join path. It needs to be enabled
using a reg key.  Implemented logic in ldap search to service file
copy requests.

Change-Id: I6b7f5e743bafc1f06ca06a2c4276fa89f66dbc0a