Added LDAP file copy functionality
Added a new LDAP control which accepts the file name/fd and blocksize and returns data read from the file.
Implemented logic to use this new functionality in the new node join path. It needs to be enabled using a reg key.
Implemented logic in ldap search to service file copy requests.

Change-Id: I9cb56917000690c743923c1c12d36e2a634c698d