Initialize OpenSSL before using VMCA client library
vmidentity used VMCA client libraries for certificate generation as part
of tenant creation flow. If concurrent tenant create requests were
received, there was a chance for VMCA_SSL_REQ_SIGN_ERR to be thrown
because openSSL library was not intialized with callback functions and
locks to make it thread safe. This change exposes the ssl init functions
in vmca client library and calls it from IDM on startup and stop of
service.

Testing:
- ran script to send create tenant requests to trigger SSL_REQ_SIGN
error. Verified that with new init changes, the error was not thrown and
the script ran successfully (./test-create-tenant.sh $domain $pass 50)
- ran all vmidentity tests

Bug ID: 2136242

Change-Id: I46ed2a0d2e458d2da084894de94044701c8e9fb9