Add security manager to sts
Adds a security manager to stsd to prevent unwanted use of classes at
runtime. The policy file specifies unwanted permissions which are
specifically denied and a whitelist of permissions needed to run sts.

On install or upgrade, it will manually extract jnidispatch lib for use.
The sts.policy file contains the enforced policy set.

Testing:
- Tested fresh install & upgrade case, then ran integration tests

Change-Id: Idebb1b7d75089c31fe66037ac4572d9f10012c9e