Change OIDC Endpoints to Server FQDN if Localhost
Before, in the OIDC client libraries, if the incoming request was from
Localhost, we would override the configured metadata endpoint with
"localhost".

Now, instead of overriding it with "localhost", we will use the server
name of the requestor if they are localhost.  With this change, if the
requestor is using the FQDN of the local host, then it would use that
value to contact the OIDC provider instead of the configured metadata
endpoint.

Testing:

    * Ensured correct URL is used through GDB when submitting REST
    request to VMDir.
    * Ensured OIDC request to VMDir never leaves localhost through PCAP

Change-Id: Ifdef674c9c8f6810936d906bc3cbaafa14c2a6a2