Add New VMAFD Client API to Create Computer OU
This change adds a new VMAFD client API to create computer account OUs.

Currently, a user can specify the OU to create a machine account under
during machine account creation or domain joining a node.  If a domain
join/machine account create operation occurs in parallel, there is a
chance for a collision during the VMDir object creation if there are
multiple Lightwave nodes in federation, and each of the object creation
calls hit different DCs.

With this change, a user can call this API first to create the parent OU
under which the machine accounts would reside.  After creating the
parent OU, then it is okay to create machine accounts--each with a
unique name--in parallel across any node in the federation.

Change-Id: I07c663263ddf40e649e65b8ff64ae1b92be21e87