Treat CreateComputerOUContainer Exists as Success
Before, when calling CreateComputerOUContainer API, if the OU to create already
exists, it would return an error.

Instead, we should swallow and ignore that error.  This change fixes
logic in VmDirLdapCreateComputerOUContainer such that if LDAP add
attempt returns with LDAP_ALREADY_EXISTS, then the error code will be
reset and treated as success.

Testing:

    * Called API to create an OU twice, saw no errors.

Change-Id: I6c6e57856e3f5a72e9cf435dd5b6c3f3add92bb1