VMCA Unit Test Infrastructure and Policy Test
This change creates a bootstrap for adding unit tests to various VMCA
functions using the CMocka framework.  In the directory vmca/unit-tests,
we mirror the directory structure for which module we want to add
unit-tests for.  The unit-tests run as a part of the build
process--calling the build/build_photon.sh script will run the tests
after Lightwave is built (make check).  If the unit-tests succeed, only
then the build will be marked as a success.

This change also refactors the VMCA service module.  To link different
structs and functions to use for unit-testing, a static library must be
made to link with the test binary.  With this refactor, the service now
has a common module which generates a static library.

Also adds an example unit-test for the VMCA policy module.
Additional unit-tests can be added by following suit.

Testing:

    * VMCA service and policy works as expected after the refactor
    * Ran Valgrind to ensure no "leaky pipes" :P
    * Ran the build script and ensure that unit-tests ran and that they
    passed

Jira ID: CNA-1234

Change-Id: I97562b7770196c4de7da7f8e2ae653f939d97a77