Refactor STS to support /isAvailable endpoint
Before, STS did not support an endpoint to indicate that it is
responding, so other services had to poll various endpoints itself to
get status. Also, STS required dependent services to be up and running
on upgrade. This change removes that dependency on upgrade and supports
an endpoint to represent overall health of the service.

Changes:
- Start health checker thread to get status of lwsm services and sts
endpoints. Start thread when servlets are initialized
- vmca is checked by getting root cert instead of running shell command
- Support /isAvailable endpoint

Testing:
- ran script to start/stop all individual services and curl endpoint,
verifying status
- verified in logs that health checker thread started/stopped correctly
- ran integration and unit tests

Change-Id: I05ce201df941a1842685f2bd043416e279591450