[POST/preprod] Coredump in post due to race condition is sasl
We observed a coredump in post preprod. We observed that in rare cases
both proxy threads call SRP bind at the same time and it causes a race
condition in cyrus sasl as sasl_client_init is not reentrant. To fix it
adding ssl_client_init along with server_init in the sasl init logic.

This will ensure that sasl_client_init is already called before the peer
threads are started.

Change-Id: I9031acfd91e29bfb3180904d7f768eaa0ba43cca