VMCA Plugin Management Interface
This change adds a generic interface to load and unload plugins for
functions which will be used to populate a VTable mapping all of the
VMCA. Plugins are DLLs which contain a well-known load and unload
available functions within it.

For example, a component of VMCA may allow plugin driven operations.
The component's initialization sequence will call `VMCAPluginInitialize`
providing a path to the .so file.  The init function will use dlopen to
load the DLL and it will then use dlsym to lookup the well known load
function within the plugin, `VMCAPluginLoad`.  The init function will
then call the plugin's load function giving the VTable for the plugin to
populate it with its functions.  The plugin handle is returned to the
component as void*.  It is the component's responsibility to cast values
obtained from the plugin interfaces to the expected type--a component
and its plugin will include a shared header which defines function
signatures and the VTable.  Once the component is finished, it can call
`VMCAPluginDeinitialize` providing the plugin handle.

Each function must have the following well-known (un)load functions:

    * `typedef DWORD (*PFN_VMCA_PLUGIN_LOAD)(VOID*);`
        * e.g. `DWORD VMCAPluginLoad(void **)`
    * `typedef VOID (*PFN_VMCA_PLUGIN_UNLOAD)(VOID);`
        * e.g. `VOID VMCAPluginUnload(void)`

A plugin's load function must populate the VTable passed in with the
function pointers defined in the shared header.

Please see the unit-tests in this commit for examples.

Testing:

    * Ran all unit tests
    * Manual plugin creation, loading, and use
    * Memcheck and FD check through Valgrind
    * Verified E2E functionality

Jira ID: CNA-2209

Change-Id: I08ddee5934a76877bb136bcbca9d26fe2eca3e99