Added ldap controls to enable/disable replication agreements
There are two controls - one to enable and one to disable.
To excercise this, we must do an ldapsearch with dn as the dn
of the replication agreement that we want to disable/enable and pass the oid of the control.
Once done, the node will stop/start pulling changes from the corresponding node in the replication agreement.

Change-Id: Ida1315c56a75f536da6c719cf25e268aef971afd