Fix DNS response code when name exists in zone
When a DNS zone has a record for a name of one type, and a query
is made for that name with a different type, VMDNS was returning
response code NXDOMAIN. The correct response in this case is to
return response code NO_ERROR with an empty answer. This change
fixes the response code.

Also emit an informational log message to report details of
the DNS query/response if request processing takes a long time.

Change-Id: I0ee93ee32da5ddc0b43f1025d2589584764a202d