CNA-2213: Added POST Schema for CA Storage
This changeset includes the schema for storing CA related attributes
persistently on POST. The attributes in the schema are derived from
the storage abstraction layer's interface and vmdir's CA schema. To
maintain consistency between lightwave and post, the additional
attributes such as cAStatus, cAEncryptedPrivateKey, and
cAEncryptedEncryptionKey are also included into vmdir as MAY attribute.

Tested the correctness of the schema by using the values picked from the
certs generated using OpenSSL.

Change-Id: I84b3dac19bdb8c1c51411f41797c4a8e7f7f9cff