CNA-2947: Util function to fetch HOTK token
With this changeset, MutentCA can now retrieve HOTK tokens for
authentication. It does so by fetching the certificate and private key
from the vecs in POST server. The commonName from the subject in the
certificate and the private key are used to get the HOTK token from the
Lightwave server.

Testing:
Along with the cmocka tests included in the changeset, retrieving a
token from a deployment of Lightwave + POST was also tested.

Change-Id: I1f69413615a45ce9bc973f25744acf9b54ee28ce