Refactor PKCS and HOTK Access Token to common
This is the first split of the ongoing work for CNA-2218 (Add POST
storage API to store Root CA with unit tests). The POST DB plugin will
make use of json utils, and HOTK Access Token util. The HOTK util uses
pkcs for parsing certificates. Since the POST plugin will be built as a
shared object, these libraries should be a part of the shared object.
Due to the previous implementation, there were challenges with
resolving the circular dependency between the libraries and
redeclaration of function definitions.

In this changeset, all the utils are moved into libcommon.

Testing:
  1. Get access token was tested against a POST deployment.
  2. make check was executed to see all the unit-tests pass.

Change-Id: I1b373c5da1c0c16399f3fbe864ce02530dfd4b01