Add HTTP functionalities to vmcommon
This is the second split of the ongoing work for CNA-2218 (Add POST
storage API to store Root CA with unit tests). The POST DB plugin will
use HTTP functionalities like setting the headers, getting the status
code, encoding the URL, and setting the body. To support that, these
functions implementations have been added to the vmcommon.

Testing:
   1. Functions were tested end-to-end with POST plugin deployment.
   2. Unit tests for the new functions have been added.

Change-Id: If90f02e13fd1b4c2aad42b41ea41fe33a2c3596e