Add POST plugin framework for DB layer
This is the third split of the ongoing work for CNA-2218 (Add POST
storage API to store Root CA with unit tests). This changeset adds the
POST plugin framework for interacting with the POST server for storage.
The DB layer will load this POST plugin to store the state
persistently.

Testing:
    1. Added unit tests to load and free the plugin from the DB layer.

Change-Id: Iab4dd74ce8cb5befa41a12ef899f2ed4f42e1a63