Support Insert CA to POST from Interface
This is the fourth and the final split of CNA-1949 (Add POST storage API
to store Root CA with unit tests). This change also includes the work
for CNA-2219 (Add POST storage API to store intermediate CA with unit
tests). The plugin uses REST calls to POST server to store the content.
It also includes the changes in the schema that reflects the changes
made in the plugin interface of the DB layer.

Testing:
 - Since this plugin is built as a shared object, cmocka unit tests
   weren't able to mock the calls (like VmHttpClient, vecs). Hence, the
   unit tests written does not make use of cmocka and only tests certain
   functions that the plugin interfaced functions will use to ensure
   the correctness of the implementation.

- Also tested inserts on a POST deployed instance.

Jira ID: CNA-1949
Jira ID: CNA-2219

Change-Id: I517fdc45907d40a6872a50311a1d3c9bbff0ee72