Support Update of Root and Intermediate CA on POST
This changeset adds the support to update Root and intermediate CA on
POST server. The update operations are "replace" only, which means, it
only updates the changes which involve change in value, and not addition
or delete of attributes. The addition and delete of values in the
attributes are not of requirement currently.

Since some members of the struct are not a pointer, which means they
can't be nullified, the caller of the plugin API needs to make sure that
the field for that entry is set correctly (eg: CA Status). The DN for
the intermediate CA is fetched by placing a GET call to the POST server,
hence, information of the Root CA is not needed to build the URI.

Testing:
- Unit tests added to validate the correctness of the request body.
- It was also tested against a POST server

Jira ID: CNA-2220
Jira ID: CNA-2216

Change-Id: I2c471bfb8ccd4e0de9caaa9108373c2a249571dd