Fix issue with replicating must to may attribute schema change
Issue: Add an entry with a must attr, apply a schema change to modify the must attr to a may attr, delte the attr from
the entry. If these changed are consumed in one update, the consumer fails to extract events because it sees attribute
metadata for the deleted must attr, but no attribute value.
Fix: Iterate over the attribute list instead of the attribute metadata.

Change-Id: Ia4750f8c6e7888630efaeec86fa8ad12c4cb9349