Update CA Schema to Include Additional Auth Data
This changeset is just an additional and optional metadata about the
tenant who created the Root/Intermediate CA. This space for metadata
could be used to store/retrieve tenant metadata for additional tasks.
The requirement of this metadata is a string. It could be json string,
or a simple string.

There were some bugs discovered (like attribute fields in response
being optional) which caused issues while parsing the response. This
changeset also resolves that alongwith.

Testing:
* Updated unit-tests to match the updated schema.
* Tested postschema patch on the POST machine to ensure schema patches
  with no errors.
* Tested save/get/update on the POST machine

Jira ID: CNA-3488

Change-Id: I34b7c9947824edd53a72f23be1d7350365a3faa1