Store Domain Join MutentCA Params in Registry
In 2fd78a16, Lightwave CLI was updated to accept parameters to enable
using a MutentCA endpoing for certificate operations, specify a MutentCA
endpiont, and specify a CA Id to use for CSR and certificate
synchronization.

This change takes those parameters and stores them in the registry
during the domain join process.

To achieve this, a new IPC domain join API--VmAfdIpcJoinVmDir3--which
takes the MutentCA endpoint and CA Id in addition to the arguments used
in the previous domain join IPC functions.

There is also some refactoring done in this change to aid a future
commit that will introduce the capability to use the provided MutentCA
endpoint and CA Id to obtain the machine SSL certificate and to
syncrhonize trusted roots from.

Testing:

    * Valgrind
    * Joining a client using the new client RPM which has domain join 3 to a
    server which has the same client version
    * Joining a client using an older client RPM which does not have domain
    join 3 to a server which has the latest client RPM that has domain join
    3
    * Ensure the registry values UseMutentCA, MutentCA Endpoint, and
    MutentCA CA Id are set correctly after domain join

Jira ID: CNA-3253

Change-Id: I573578fdc2dd30565e34be05942bd21657797b9f