Lock Manager for MutentCA POST Plugin
This changeset implements the lock manager for DB layer. The
implementation makes use of 2 additional attributes for locking:
resourceLockOwner and resourceLockExpirationTime. The lock owner is the
identifier for the lock and expiration time is used to handle the time
drift that could occur between systems. If the values are 'noOwner' and
0 respectively, it means the DN is not locked. For other values, the
lock has been held. This lock can be released if it is the same owner
that wants to unlock or if the expiration time has exceeded the current
time.

Since there are no delete operations currently in MutentCA's db layer,
the ADD operations can be made without a lock. Locks are required when
making an update in the existing entries, to avoid 2 simultaneous updates
on the same DN.

The implementation is similar to spin-lock, where if the condition is
not satisfied, the requesting thread goes to sleep (for 500ms). Also,
such condition checks will occur for a maximum of 1000 attempts. And, the
maximum clock time drift acceptable for the lock manager is 60 seconds.

Testing:
- Tested patching of schema to make sure there are no errors.
- Updated the REST request bodies to match the expected JSON request.
- Added unit test for testing the REST request body for lock/unlock.
- Performed some extent of concurrency test by adding dummy values to
  resourceLockOwner and resourceLockExpirationTime to a given DN, and
  then validating the sleep for 500ms, getting access to lock after
  exceeding the expiration time, and ensuring that lock bails out with
  error once it exceed its maximum attempts.

Jira ID: CNA-3491

Change-Id: I732d4d9f29deb86c22d10fb71596f1dd40452e21